Organise tests and add lib to gitignore
