+++ .gitignore
@@ -6,3 +6,4 @@ coverage.json
 .DS_Store
 *.swp
 *.log
+lib/
