--- lib/index.d.ts
@@ -1,5 +1,0 @@
-export declare const log: {
-    onReducerCalled(state: any, prev: any, name: any, ...args: any[]): void;
-    onEffectCalled(state: any, name: any, ...args: any[]): void;
-};
-export default function (configuration: any): any;
